Fix jack-o'-lantern orientation on placement. Closes GH-478
Move the Pumpkin interface facing logic up to a BlockPumpkinBase class
shared by both BlockPumpkin and BlockJackOLantern. Placed
jack-o'-lanterns now face the player, just as pumpkins face the player.
