+++ src/main/java/net/glowstone/block/blocktype/BlockJackOLantern.java
@@ -1,0 +1,10 @@
+package net.glowstone.block.blocktype;
+
+import static org.bukkit.Material.*;
+
+public class BlockJackOLantern extends BlockPumpkinBase {
+
+    public BlockJackOLantern() {
+        super(JACK_O_LANTERN);
+    }
+}
